Fix mixin type in GreenWallet

+++ src/GreenWallet/Tools.cpp
@@ -151,7 +151,7 @@ namespace Tools {
     );
   }
 
-  uint64_t getScanHeight()
+  uint32_t getScanHeight()
   {
     while (true)
     {
